transmart-oauth refactoring
- move spring beans definitions
- change transmart-server gsp files
