+++ transmart-server/build.gradle
@@ -76,7 +76,7 @@ dependencies {
 }
 
 bootRun {
-    jvmArgs('-Dspring.output.ansi.enabled=always')
+    jvmArgs('-Xmx4096m')
 }
