+++ transmart-server/grails-app/views/login/denied.gsp
@@ -1,0 +1,16 @@
+<head>
+    <meta name='layout' content='main'/>
+    <title>Authentication Error in ${grailsApplication.config.com.recomdata.appTitle}</title>
+</head>
+
+<body>
+<div style="text-align: center;">
+    <div style="width: 600px; margin: 50px auto 50px auto;">
+        <img style="display: block; margin: 12px auto;"
+             src="${resource(dir: 'images', file: grailsApplication.config.com.recomdata.largeLogo)}" alt="Transmart"/>
+        <div style="text-align: center;">If you would like to use the tranSMART system, please contact a <a
+                href="mailto:${grailsApplication.config.com.recomdata.administrator}??subject=tranSMART new user request&body=Please, add me as a new user to the tranSMART system.">tranSMART Admin</a>
+        </div>
+    </div>
+</div>
+</body>
